requirements.txt
pyyaml
regex
argparse
fastapi
python-multipart
uvicorn
prisma
python-magic
orjson